Replace 0.3.5 of `backgrid-filter` with 0.3.7
This fixes issue #7 from `wrensec-deps`. It looks like ForgeRock had/has a custom version of `backgrid-filter-0.3.5` that includes changes from 0.3.6. This fixes the confusion by dropping 0.3.5 entirely and just using 0.3.7. This is a breaking change for Wren projects that currently are depending on 0.3.5 -- they will need to be updated to 0.3.7.